fix: nextStartWeight to Long

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/schemas/v2/GroupBoardArticlePaginationBase.kt
@@ -15,5 +15,5 @@ data class GroupBoardArticlePaginationBase(
     @SerializedName("hasNext")
     var hasNext: Boolean?,
     @SerializedName("nextStartWeight")
-    var nextStartWeight: Int?
+    var nextStartWeight: Long?
 )
